requirements.txt
# AetherContracts — Python Dependencies (Docker)
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.9.0
pyyaml==6.0.2
requests==2.32.0
huggingface-hub==0.25.0
psutil==6.0.0
